Cleanup
Used to delete old backups using the YYYYMMDD date format as part of the file name when created. It deletes the prior month, but can be easily modified to delete prior year files if you want to keep them around longer. WARNING: Once January rolls around it will not delete December files in the current version, I'm still working on that.