Basic install and config for Pihole on a Pi
This will update your Pi, install and configure UFW, and install Pi Hole. It also loads several blocklists as a starting point for you. You will need to curate them for your own personal preferences, you can remove them or add more via the web interface. Write down the admin password Pi Hole shows you after install, you will need it to log into the web interface. To fully configure Pihole as your DNS provider you will need to follow the instructions for the top level router in your home lab, it varies by brand and model.

If you are self-hosting a web site on prem DO NOT USE THAT MACHINE FOR YOUR DNS SERVER. The web server and DNS servers must be different machines. I do not know if this is Pi Hole specific or a general thing but your website will not be visible from outside your network if it's on the same machine as Pi Hole.